Restructure v2 task files
There is no change in functionality but instead, merged some files, and renamed to shorter names. This resembles dependabot-action to some extent. Hopefully, it does not look as complex to other who want to contribute.
