+++ extension/tasks/dependabotV2/azure-devops/models.ts
@@ -1,5 +1,15 @@
 import { GitPullRequestMergeStrategy } from 'azure-devops-node-api/interfaces/GitInterfaces';
-import { IFileChange } from './IFileChange';
+import { VersionControlChangeType } from 'azure-devops-node-api/interfaces/TfvcInterfaces';
+
+/**
+ * File change
+ */
+export interface IFileChange {
+  changeType: VersionControlChangeType;
+  path: string;
+  content: string;
+  encoding: string;
+}
 
 /**
  * Pull request properties
