--- extension/tasks/dependabotV2/utils/azure-devops/interfaces/IFileChange.ts
@@ -1,11 +1,0 @@
-import { VersionControlChangeType } from 'azure-devops-node-api/interfaces/TfvcInterfaces';
-
-/**
- * File change
- */
-export interface IFileChange {
-  changeType: VersionControlChangeType;
-  path: string;
-  content: string;
-  encoding: string;
-}
