--- extension/tasks/dependabotV2/utils/dependabot-cli/interfaces/IDependabotUpdateOperation.ts
@@ -1,9 +1,0 @@
-import { IDependabotUpdate } from '../../dependabot/interfaces/IDependabotConfig';
-import { IDependabotUpdateJobConfig } from './IDependabotUpdateJobConfig';
-
-/**
- * Represents a single Dependabot CLI update operation
- */
-export interface IDependabotUpdateOperation extends IDependabotUpdateJobConfig {
-  config: IDependabotUpdate;
-}
