--- extension/tasks/dependabotV2/utils/dependabot-cli/interfaces/IDependabotUpdateOperationResult.ts
@@ -1,11 +1,0 @@
-/**
- * Represents the output of a Dependabot CLI update operation
- */
-export interface IDependabotUpdateOperationResult {
-  success: boolean;
-  error: Error;
-  output: {
-    type: string;
-    data: any;
-  };
-}
