--- extension/tasks/dependabotV2/utils/github/IPackage.ts
@@ -1,4 +1,0 @@
-export interface IPackage {
-  name: string;
-  version?: string;
-}
