--- extension/tasks/dependabotV2/utils/github/ISecurityVulnerability.ts
@@ -1,11 +1,0 @@
-import { IPackage } from './IPackage';
-import { ISecurityAdvisory } from './ISecurityAdvisory';
-import { PackageEcosystem } from './PackageEcosystem';
-
-export interface ISecurityVulnerability {
-  ecosystem: PackageEcosystem;
-  package: IPackage;
-  advisory: ISecurityAdvisory;
-  vulnerableVersionRange: string;
-  firstPatchedVersion: string;
-}
