+++ extension/tasks/dependabotV2/utils/placeholder.ts
