+++ extension/tasks/dependabotV2/utils/shared-variables.ts
@@ -1,10 +1,7 @@
 import * as tl from 'azure-pipelines-task-lib/task';
-import { DEFAULT_EXPERIMENTS } from './dependabot/experiments';
-import extractHostname from './extractHostname';
-import extractOrganization from './extractOrganization';
-import extractVirtualDirectory from './extractVirtualDirectory';
-import getAzureDevOpsAccessToken from './getAzureDevOpsAccessToken';
-import getGithubAccessToken from './getGithubAccessToken';
+import { DEFAULT_EXPERIMENTS } from '../dependabot/experiments';
+import { getAzureDevOpsAccessToken, getGithubAccessToken } from './tokens';
+import { extractHostname, extractOrganization, extractVirtualDirectory } from './url-parts';
 
 export interface ISharedVariables {
   /** URL of the organization. This may lack the project name */
